chore: add TODO comments for remaining CI-weakened tests
- nix checkPhase: flow integration tests (Python, Ruby, Noir) skipped
  because recorders/compilers aren't in the build sandbox
- Noir flow test: silently passes when nargo is missing
- xvfb-run workaround: temp-file hack for nix xvfb-run cleanup bug
- Fix pre-existing clippy lint (unnecessary i64 cast in flow_preloader)

+++ src/db-backend/src/flow_preloader.rs
@@ -325,7 +325,7 @@ impl<'a> CallFlowPreloader<'a> {
             // the function entry point.
             if step_id.0 == 0 && self.mode == FlowMode::Call && self.location.line > 0 {
                 replay.jump_to(StepId(0))?;
-                let bp = replay.add_breakpoint(&self.location.path, self.location.line as i64)?;
+                let bp = replay.add_breakpoint(&self.location.path, self.location.line)?;
                 let hit = replay.step(Action::Continue, true)?;
                 let _ = replay.delete_breakpoint(&bp);
                 if hit {
